✨🔧 (CleanupDataFiles.ps1, ImagesToBlobStorage.ps1): enhance logging and file handling
Enhance the scripts to provide more informative logging and handle cases where no files are found. The changes include:
- Adding checks for the presence of files before attempting deletion, logging a message if no files are found.
- Calculating and logging the total size of files to be deleted, providing better insight into the operation.
- Replacing `Write-Debug` with `Write-DebugLog` and `Write-ErrorLog` for consistency and clarity in logging.

These improvements aim to make the scripts more robust and user-friendly by providing detailed feedback during execution.

+++ .powershell/_includes/CleanupDataFiles.ps1
@@ -8,7 +8,15 @@ function Delete-LocalDataFiles {
     $count = 0
     try {
         Write-InfoLog "Deleting all data file files locally from '$LocalPath'..."
-        Get-ChildItem -Path $env:LOCAL_IMAGE_PATH -Recurse -Include data.captions.*.srt, data.captions.json, data.json, data.index.classifications.json  | ForEach-Object {
+        $files = Get-ChildItem -Path $env:LOCAL_IMAGE_PATH -Recurse -Include data.captions.*.srt, data.captions.json, data.json, data.index.classifications.json
+        if ($files.Count -eq 0) {
+            Write-InfoLog "No files found."
+            return 0;
+        }
+        $size = ($files | Measure-Object -Property Length -Sum).Sum 
+        $sizeString = "{0:N2} MB" -f ($size / 1MB)
+        Write-InfoLog "Found ($($files.Count)) files totalling $sizeString."
+        $files | ForEach-Object {
             try {
                 $count++
                 Remove-Item -Path $_.FullName -Force
